+++ packages/db/help/migrations create.txt
@@ -1,0 +1,1 @@
+Create next migration file
