+++ packages/db/test/migrate/sqlite3.test.mjs
@@ -3,7 +3,7 @@ import path from 'path'
 import pino from 'pino'
 import split from 'split2'
 import sqlite from '@databases/sqlite'
-import { execute } from '../../lib/migrator.mjs'
+import { execute } from '../../lib/migrate.mjs'
 import ConfigManager from '../../lib/config.js'
 import { cleanSQLite } from './helper.mjs'
 import { urlDirname } from '../../lib/utils.js'
